+++ CHANGELOG.md
@@ -28,6 +28,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 - [MINOR] Added `async_util.gather_batched` function to `async_util`
 - [MINOR] Added more exception classes
 - [MAJOR] Added `apiPathPattern` to `logging.api`
+- [MAJOR] Added `chainId` to `EthClientInterface`
+- [MINOR] Added `call` and `multicall` to `EthClientInterface`
 
 ### Changed